backend/enhancement: Auto-disable Fleet Owner or Driver upon removal of mandatory document
