+++ Backend/app/dashboard/CommonAPIs/spec/ProviderPlatform/Management/API/DriverRegistration.yaml
@@ -158,12 +158,15 @@ apis:
             query:
               - requestorId: Text
             response:
-              type: APISuccess
+              type: UnlinkDocumentResp
 
 types:
   ValidationStatus:
     enum: "APPROVED,DECLINED,AUTO_APPROVED,AUTO_DECLINED,NEEDS_REVIEW"
     derive: "Read"
+  UnlinkDocumentResp:
+   - mandatoryDocumentRemoved : Bool
+   - recordType: NewType
   DLDetails:
     - driverName: Maybe Text
     - driverLicenseNumber: Text
