+++ Backend/app/dashboard/Lib/src/Domain/Action/Dashboard/Person.hs
@@ -548,7 +548,8 @@ buildPerson req dashboardAccessType = do
         verified = Nothing,
         rejectionReason = Nothing,
         rejectedAt = Nothing,
-        passwordUpdatedAt = Just now
+        passwordUpdatedAt = Just now,
+        approvedBy = Nothing
       }
 
 data UpdatePersonReq = UpdatePersonReq
