+++ Backend/app/dashboard/provider-dashboard/src/Domain/Action/ProviderPlatform/Management/DriverRegistration.hs
@@ -36,12 +36,13 @@ import qualified "lib-dashboard" Domain.Types.Person as DP
 import qualified Domain.Types.Transaction as DT
 import "lib-dashboard" Environment
 import Kernel.Prelude
-import Kernel.Types.APISuccess (APISuccess)
+import Kernel.Types.APISuccess (APISuccess (Success))
 import Kernel.Types.Beckn.City as City
 import Kernel.Types.Id
 import Kernel.Utils.Common
 import qualified SharedLogic.Transaction as T
 import Storage.Beam.CommonInstances ()
+import qualified "lib-dashboard" Storage.Queries.Person as QP
 import "lib-dashboard" Tools.Auth
 import "lib-dashboard" Tools.Auth.Merchant
 
@@ -150,6 +151,10 @@ postDriverRegistrationRegisterAadhaar merchantShortId opCity apiTokenInfo driver
 postDriverRegistrationUnlinkDocument :: ShortId DM.Merchant -> City.City -> ApiTokenInfo -> Id Common.Driver -> Common.DocumentType -> Flow APISuccess
 postDriverRegistrationUnlinkDocument merchantShortId opCity apiTokenInfo personId documentType = do
   checkedMerchantId <- merchantCityAccessCheck merchantShortId apiTokenInfo.merchant.shortId opCity apiTokenInfo.city
-  transaction <- buildTransaction apiTokenInfo (Just personId) (Just documentType)
   let mbRequestorId = determineRequestorId apiTokenInfo personId
-  T.withTransactionStoring transaction $ Client.callManagementAPI checkedMerchantId opCity (.driverRegistrationDSL.postDriverRegistrationUnlinkDocument) personId documentType mbRequestorId
+  transaction <- buildTransaction apiTokenInfo (Just personId) (Just documentType)
+  T.withTransactionStoring transaction $ do
+    res <- Client.callManagementAPI checkedMerchantId opCity (.driverRegistrationDSL.postDriverRegistrationUnlinkDocument) personId documentType mbRequestorId
+    when res.mandatoryDocumentRemoved $ do
+      QP.updatePersonVerifiedStatus (cast personId) False
+    pure Success
